@@ -5,7 +5,7 @@ repository: https://github.com/baseflow/flutter-geocoding/tree/main/geocoding
 issue_tracker: https://github.com/Baseflow/flutter-geocoding/issues
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
+  sdk: ">=3.3.0 <4.0.0"
   flutter: ">=3.0.0"
 
 dependencies:
@@ -20,7 +20,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^3.0.1
+  flutter_lints: ^5.0.0
   mockito: ^5.0.0
   plugin_platform_interface: ^2.0.0
   
